requirements.txt
Flask==0.10.1
Flask-Sockets==0.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
argparse==1.2.1
gevent==0.13.8
gevent-socketio==0.3.6
gevent-websocket==0.3.6
greenlet==0.4.1
gunicorn==18.0
itsdangerous==0.23
redis==2.8.0
wsgiref==0.1.2


#Flask==0.9
#Flask-SQLAlchemy==2.0
#Jinja2==2.7.3
#MarkupSafe==0.23
#SQLAlchemy==0.9.8
#Werkzeug==0.9.6
#argparse==1.2.1
#gevent==1.0.1
#gevent-socketio==0.3.5-rc2
#gevent-websocket==0.9.3
#greenlet==0.4.5
#redis==2.10.3
#wsgiref==0.1.2
